requirements.txt
pyyaml>=6.0
pandas>=2.0
huggingface-hub>=0.20
